further work on user service and model

+++ spring-boot-example/src/main/java/ftn/isa/model/AppointmentReview.java
@@ -2,16 +2,16 @@ package ftn.isa.model;
 
 public class AppointmentReview 
 {
-	private int id;
+	private Long id;
 	private Appointment appointment;
 	private String description;
 	
-	public int getId() 
+	public Long getId() 
 	{
 		return id;
 	}
 	
-	public void setId(int id) 
+	public void setId(Long id) 
 	{
 		this.id = id;
 	}
