+++ spring-boot-example/src/main/java/ftn/isa/model/Complaint.java
@@ -2,14 +2,16 @@ package ftn.isa.model;
 
 public class Complaint 
 {
-	private int id;
+	private Long id;
 	private String description;
 	private MedicalCenter medicalCenter;
 	private MedicalStaff medicalStaf;
-	public int getId() {
+	
+	public Long getId() 
+	{
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getDescription() {
